Add safeguards for .home_cleanup script

+++ .home_cleanup.sh
@@ -2,11 +2,18 @@
 # Clean up workspace
 # shellcheck disable=SC2016
 
-echo   '# This will clean up the workspace, ALL WILL BE LOST!'
-printf '# ... Are you sure? If so, type `yes`: '
+if [[ "$(git rev-parse --show-toplevel)" = "$HOME" ]]; then
+	>&2 echo '# This is the $HOME repo, DANGER! Exiting ...'
+	exit 1
+fi
+
+>&2 echo   '# This will clean up the workspace, ALL WILL BE LOST!'
+>&2 printf '# ... Are you sure? If so, type `yes`: '
 
 read -r affirm
 if [[ "$affirm" = 'yes' ]]; then
 	git checkout -- .
 	git clean -dfx
 fi
+
+# vim: set noexpandtab:
